+++ src/index.css
@@ -4,6 +4,7 @@
 
 body {
   margin: 0;
+  @apply bg-gray-100 dark:bg-gray-900;
   font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', 'Roboto', 'Oxygen',
     'Ubuntu', 'Cantarell', 'Fira Sans', 'Droid Sans', 'Helvetica Neue',
     sans-serif;
